+++ src/logonserver/Server/Master.cpp
@@ -17,7 +17,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include <Network/Network.h>
 #include "Console/LogonConsole.h"
 #include "LogonConf.hpp"
-#include <Util/Strings.hpp>
+#include <Utilities/Strings.hpp>
 
 using std::chrono::milliseconds;
