+++ src/shared/Util.cpp
@@ -4,7 +4,7 @@ This file is released under the MIT license. See README-MIT for more information
 */
 
 #include "Util.hpp"
-#include "Util/Strings.hpp"
+#include "Utilities/Strings.hpp"
 #include <iostream>
 #include <vector>
 #include <string>
