+++ src/shared/Utilities/Strings.cpp
