+++ src/shared/Utilities/Strings.hpp
