+++ src/world/Chat/ChannelMgr.cpp
@@ -12,7 +12,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include "Server/World.h"
 #include "Server/WorldSession.h"
 #include "Server/Packets/SmsgChannelNotify.h"
-#include "Util/Strings.hpp"
+#include "Utilities/Strings.hpp"
 
 using namespace AscEmu::Packets;
