+++ src/world/Chat/Commands/LookupCommands.cpp
@@ -8,7 +8,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include "Storage/MySQLDataStore.hpp"
 #include "Storage/MySQLStructures.h"
 #include "Spell/SpellMgr.hpp"
-#include "Util/Strings.hpp"
+#include "Utilities/Strings.hpp"
 
 //.lookup achievement
 //////////////////////////////////////////////////////////////////////////////////////////
