+++ src/world/Management/AuctionHouse.cpp
@@ -16,7 +16,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include "Server/Packets/SmsgAuctionBidderListResult.h"
 #include "Server/Packets/SmsgAuctionListResult.h"
 #include "Server/Packets/CmsgAuctionListItems.h"
-#include "Util/Strings.hpp"
+#include "Utilities/Strings.hpp"
 
 using namespace AscEmu::Packets;
