+++ src/world/Management/Guild/GuildMgr.cpp
@@ -9,7 +9,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include "Guild.hpp"
 #include "Management/ObjectMgr.h"
 #include "Server/MainServerDefines.h"
-#include "Util/Strings.hpp"
+#include "Utilities/Strings.hpp"
 
 GuildMgr& GuildMgr::getInstance()
 {
