+++ src/world/Map/Maps/InstanceMgr.cpp
@@ -10,7 +10,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include "Map/Management/MapMgr.hpp"
 #include "Server/Definitions.h"
 #include "Server/Script/ScriptMgr.h"
-#include "Util/Strings.hpp"
+#include "Utilities/Strings.hpp"
 
 InstanceSaved::InstanceSaved(uint32_t mapId, uint32_t instanceId, InstanceDifficulty::Difficulties difficulty, time_t resetTime, bool canReset)
     : m_resetTime(resetTime),
