+++ src/world/Objects/Units/Players/Player.cpp
@@ -128,7 +128,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include "Server/Warden/SpeedDetector.h"
 #include "Spell/Definitions/SpellEffects.hpp"
 #include "Storage/WorldStrings.h"
-#include "Util/Strings.hpp"
+#include "Utilities/Strings.hpp"
 
 using namespace AscEmu::Packets;
 using namespace MapManagement::AreaManagement;
